Extend Home Assistant API REST client to support types for "Backup" integration
Extended the Home Assistant REST client implemented in GH-3 [3] to
support types for the (built-in) "Backup" integration [1], enabling
programmatic parsing and processing of the `backup.json` file [2] that
is included in the archive files created by the integration. This allows
for processing the archive files in custom backup automations, e.g.
to extract only specific data.

> Goals & Objectives

- Provide simple typed structs to define the content of the
  `backup.json` file.

[1]: https://www.home-assistant.io/integrations/backup
[2]: https://github.com/home-assistant/core/blob/e572f8d4/homeassistant/components/backup/util.py#L74-L76
[3]: https://github.com/svengreb/home-ops-cli/issues/3

GH-5

Signed-off-by: Sven Greb <development@svengreb.de>