Build and release process with "ko" and automated container releases to GHCR via GitHub Actions
Set up containerized [2] builds of the CLI using ko [1], and automate
image publishing to the GitHub Container Registry (GHCR) [3] through
GitHub Actions [4]. This enables consistent, reproducible images and
streamlined releases.
The configuration for ko [4] is persisted through an `.ko.yaml` file in
the root of the repository which is picked up during the build process.
The `ko-build/setup-ko` GitHub Action [6] is used to run the whole
process in an automated fashion for development states, requested
explicitly though repository maintainers, and tagged releases.

> Goals & Objectives

- Add a `.ko.yaml` configuration file for building the CLI as a
  container image.
- Configure ko to publish images to `ghcr.io/svengreb/home-ops-cli`.
- Create GitHub Actions workflows to:
  - Build and push images on each merge to the `main` branch.
  - Build and publish versioned images when a Git tag is created.
- Ensure images include proper labels, metadata, and version
  information.
- Ensure SBOMs are generated.

[1]: https://ko.build
[2]: https://opencontainers.org
[3]: https://docs.github.com/en/packages/working-with-a-github-packages-registry/working-with-the-container-registry
[4]: https://github.com/features/actions
[5]: https://ko.build/configuration
[6]: https://github.com/ko-build/setup-ko

GH-8

Signed-off-by: Sven Greb <development@svengreb.de>